refactor: convert growth parameters to float-based percentages
- Change growth parameters from integer-based to float-based for more precise control\n- Update growth calculation logic in both pool and fast path\n- Modify default values to use percentages instead of multipliers\n- Update validation and builder methods to handle float values\n- Improve type safety and precision in growth calculations

+++ pool/pool_config_builder.go
@@ -60,21 +60,21 @@ func (b *poolConfigBuilder) SetHardLimit(count int) PoolConfigBuilder {
 // SetGrowthExponentialThresholdFactor sets the threshold factor for switching from
 // exponential to fixed growth. When the pool's capacity exceeds this threshold,
 // growth switches from percentage-based to fixed-step growth.
-func (b *poolConfigBuilder) SetGrowthExponentialThresholdFactor(factor int) PoolConfigBuilder {
+func (b *poolConfigBuilder) SetGrowthExponentialThresholdFactor(factor float64) PoolConfigBuilder {
 	b.config.growth.thresholdFactor = factor
 	return b
 }
 
 // SetGrowthFactor sets the growth factor used in exponential growth mode.
 // This determines how much the pool grows by factor when below the exponential threshold.
-func (b *poolConfigBuilder) SetGrowthFactor(factor int) PoolConfigBuilder {
+func (b *poolConfigBuilder) SetGrowthFactor(factor float64) PoolConfigBuilder {
 	b.config.growth.bigGrowthFactor = factor
 	return b
 }
 
 // SetFixedGrowthFactor sets the fixed growth factor used after exponential growth ends.
 // This determines the fixed step size for growth when above the exponential threshold.
-func (b *poolConfigBuilder) SetFixedGrowthFactor(factor int) PoolConfigBuilder {
+func (b *poolConfigBuilder) SetFixedGrowthFactor(factor float64) PoolConfigBuilder {
 	b.config.growth.controlledGrowthFactor = factor
 	return b
 }
@@ -165,21 +165,21 @@ func (b *poolConfigBuilder) SetFastPathGrowthEventsTrigger(count int) PoolConfig
 
 // SetFastPathGrowthPercent sets the growth percentage for the fast path.
 // This determines how much the L1 cache grows by percentage.
-func (b *poolConfigBuilder) SetFastPathGrowthFactor(factor int) PoolConfigBuilder {
+func (b *poolConfigBuilder) SetFastPathGrowthFactor(factor float64) PoolConfigBuilder {
 	b.config.fastPath.growth.bigGrowthFactor = factor
 	return b
 }
 
 // SetFastPathExponentialThresholdFactor sets the exponential threshold factor for fast path growth.
 // This determines when the L1 cache switches from exponential to fixed growth.
-func (b *poolConfigBuilder) SetFastPathExponentialThresholdFactor(percent int) PoolConfigBuilder {
+func (b *poolConfigBuilder) SetFastPathExponentialThresholdFactor(percent float64) PoolConfigBuilder {
 	b.config.fastPath.growth.thresholdFactor = percent
 	return b
 }
 
 // SetFastPathFixedGrowthFactor sets the fixed growth factor for fast path growth.
 // This determines the fixed step size for L1 cache growth above the threshold.
-func (b *poolConfigBuilder) SetFastPathFixedGrowthFactor(factor int) PoolConfigBuilder {
+func (b *poolConfigBuilder) SetFastPathFixedGrowthFactor(factor float64) PoolConfigBuilder {
 	b.config.fastPath.growth.controlledGrowthFactor = factor
 	return b
 }
